+++ browser/Makefile
@@ -1,0 +1,3 @@
+run:
+	GOOS=js GOARCH=wasm go build -o assets/tun.wasm ./wasm/
+	go run main.go
