+++ browser/assets/index.html
@@ -28,7 +28,6 @@
     <label for="peerAllowedIPs">Wireguard Peer AllowedIPs:</label>
     <input id="peerAllowedIPs" type=input size="50" value="Paste other peer AllowedIPs">
     <p/>
-    <!--<input type=button value="wgPrivateKey" onclick='wgPrivateKey(prompt("Wireguard Private Key", ""))'>-->
     <input type=button value="start" onclick='connect()'>
 </form>
 </body>
